+++ NEWS.md
@@ -1,8 +1,9 @@
-# CHANGES IN radiant.basics VERSION 0.7.1
+# CHANGES IN radiant.basics VERSION 0.7.3
 
 ## NEW FEATURES
 
 - Allow sampling data for correlation plot. Default is 1,000 obs
+- Show df name in output
 
 ## BUG FIXES